# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SleepCreate": {
        "properties": {
          "baby_id": {
            "title": "Baby Id",
            "type": "integer"
          },
          "end_time": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Time"
          },
          "notes": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "start_time": {
            "format": "date-time",
            "title": "Start Time",
            "type": "string"
          }
        },
        "required": [
          "baby_id",
          "start_time"
        ],
        "title": "SleepCreate",
        "type": "object"
      },
      "SleepResponse": {
        "properties": {
          "baby_id": {
            "title": "Baby Id",
            "type": "integer"
          },
          "comment_count": {
            "default": 0,
            "title": "Comment Count",
            "type": "integer"
          },
          "end_time": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Time"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "notes": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "recorded_by_display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recorded By Display Name"
          },
          "start_time": {
            "format": "date-time",
            "title": "Start Time",
            "type": "string"
          },
          "unlocked_achievements": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/UnlockedAchievementInfo"
            },
            "title": "Unlocked Achievements",
            "type": "array"
          },
          "user_id": {
            "title": "User Id",
            "type": "integer"
          }
        },
        "required": [
          "baby_id",
          "start_time",
          "id",
          "user_id"
        ],
        "title": "SleepResponse",
        "type": "object"
      },
      "SleepUpdate": {
        "properties": {
          "end_time": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Time"
          },
          "notes": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "start_time": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Time"
          }
        },
        "title": "SleepUpdate",
        "type": "object"
      },
...
      "TemperatureCreate": {
        "properties": {
          "baby_id": {
            "title": "Baby Id",
            "type": "integer"
          },
          "measured_at": {
            "format": "date-time",
            "title": "Measured At",
            "type": "string"
          },
          "method": {
            "$ref": "#/components/schemas/TemperatureMethod",
            "default": "AXILLARY"
          },
          "notes": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "temperature": {
            "maximum": 42.0,
            "minimum": 34.0,
            "title": "Temperature",
            "type": "number"
          }
        },
        "required": [
          "baby_id",
          "measured_at",
          "temperature"
        ],
        "title": "TemperatureCreate",
        "type": "object"
      },
      "TemperatureMethod": {
        "enum": [
          "AXILLARY",
          "EAR",
          "FOREHEAD",
          "RECTAL"
        ],
        "title": "TemperatureMethod",
        "type": "string"
      },
      "TemperatureResponse": {
        "properties": {
          "baby_id": {
            "title": "Baby Id",
            "type": "integer"
          },
          "comment_count": {
            "default": 0,
            "title": "Comment Count",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "measured_at": {
            "format": "date-time",
            "title": "Measured At",
            "type": "string"
          },
          "method": {
            "$ref": "#/components/schemas/TemperatureMethod",
            "default": "AXILLARY"
          },
          "notes": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "recorded_by_display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recorded By Display Name"
          },
          "temperature": {
            "maximum": 42.0,
            "minimum": 34.0,
            "title": "Temperature",
            "type": "number"
          },
          "unlocked_achievements": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/UnlockedAchievementInfo"
            },
            "title": "Unlocked Achievements",
            "type": "array"
          },
          "user_id": {
            "title": "User Id",
            "type": "integer"
          }
        },
        "required": [
          "baby_id",
          "measured_at",
          "temperature",
          "id",
          "user_id"
        ],
        "title": "TemperatureResponse",
        "type": "object"
      },
      "TemperatureUpdate": {
        "properties": {
          "measured_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Measured At"
          },
          "method": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TemperatureMethod"
              },
              {
                "type": "null"
              }
            ]
          },
          "notes": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "temperature": {
            "anyOf": [
              {
                "maximum": 42.0,
                "minimum": 34.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Temperature"
          }
        },
        "title": "TemperatureUpdate",
        "type": "object"
      },
...
      "UnlockedAchievementInfo": {
        "description": "記録APIレスポンスに含める実績解除情報",
        "properties": {
          "achievement_id": {
            "title": "Achievement Id",
            "type": "string"
          },
          "category": {
            "title": "Category",
            "type": "string"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "icon": {
            "title": "Icon",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "rarity": {
            "title": "Rarity",
            "type": "string"
          }
        },
        "required": [
          "achievement_id",
          "name",
          "icon",
          "category",
          "rarity",
          "description"
        ],
        "title": "UnlockedAchievementInfo",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Botoro API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/sleeps/": {
      "get": {
        "operationId": "get_sleeps_api_sleeps__get",
        "parameters": [
          {
            "in": "query",
            "name": "baby_id",
            "required": true,
            "schema": {
              "title": "Baby Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SleepResponse"
                  },
                  "title": "Response Get Sleeps Api Sleeps  Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Sleeps",
        "tags": [
          "sleeps"
        ]
      },
      "post": {
        "operationId": "create_sleep_api_sleeps__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SleepCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SleepResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Sleep",
        "tags": [
          "sleeps"
        ]
      }
    },
    "/api/sleeps/{sleep_id}": {
      "delete": {
        "operationId": "delete_sleep_api_sleeps__sleep_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "sleep_id",
            "required": true,
            "schema": {
              "title": "Sleep Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Sleep",
        "tags": [
          "sleeps"
        ]
      },
      "patch": {
        "operationId": "update_sleep_api_sleeps__sleep_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "sleep_id",
            "required": true,
            "schema": {
              "title": "Sleep Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SleepUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SleepResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Sleep",
        "tags": [
          "sleeps"
        ]
      }
    },
    "/api/temperatures/": {
      "get": {
        "operationId": "get_temperatures_api_temperatures__get",
        "parameters": [
          {
            "in": "query",
            "name": "baby_id",
            "required": true,
            "schema": {
              "title": "Baby Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "skip",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Skip",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TemperatureResponse"
                  },
                  "title": "Response Get Temperatures Api Temperatures  Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Temperatures",
        "tags": [
          "temperatures"
        ]
      },
      "post": {
        "operationId": "create_temperature_api_temperatures__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TemperatureCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemperatureResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Temperature",
        "tags": [
          "temperatures"
        ]
      }
    },
    "/api/temperatures/{record_id}": {
      "delete": {
        "operationId": "delete_temperature_api_temperatures__record_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "record_id",
            "required": true,
            "schema": {
              "title": "Record Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Temperature",
        "tags": [
          "temperatures"
        ]
      },
      "put": {
        "operationId": "update_temperature_api_temperatures__record_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "record_id",
            "required": true,
            "schema": {
              "title": "Record Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TemperatureUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemperatureResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Temperature",
        "tags": [
          "temperatures"
        ]
      }
    },
...
  }
}